+++ cmd/utils/flags.go
@@ -2419,7 +2419,7 @@ func MakeChain(ctx *cli.Context, stack *node.Node, readonly bool) (*core.BlockCh
 	walDir := stack.Config().GolemBaseWriteAheadLogDir
 	if walDir != "" {
 		chain, err := core.NewBlockChainWithOnNewBlock(chainDb, cache, gspec, nil, engine, vmcfg, nil, func(block *types.Block, receipts []*types.Receipt) error {
-			return wal.WriteLogForBlock(walDir, block, receipts)
+			return wal.WriteLogForBlock(walDir, block, config.ChainID, receipts)
 		})
 		if err != nil {
 			Fatalf("Can't create BlockChain with onNewBlock: %v", err)