+++ golem-base/storageutil/active_payload.go
@@ -1,5 +1,7 @@
 package storageutil
 
+import "github.com/ethereum/go-ethereum/common"
+
 //go:generate go run ../../rlp/rlpgen -type ActivePayload -out gen_active_payload_rlp.go
 
 // ActivePayload represents a payload that is currently active in the storage layer.
@@ -8,10 +10,11 @@ package storageutil
 // The Key of the entity is derived from the payload content and the transaction hash where the entity was created.
 
 type ActivePayload struct {
-	ExpiresAtBlock     uint64
-	Payload            []byte
-	StringAnnotations  []StringAnnotation
-	NumericAnnotations []NumericAnnotation
+	ExpiresAtBlock     uint64              `json:"expiresAtBlock"`
+	Payload            []byte              `json:"payload"`
+	StringAnnotations  []StringAnnotation  `json:"stringAnnotations"`
+	NumericAnnotations []NumericAnnotation `json:"numericAnnotations"`
+	Owner              common.Address      `json:"owner"`
 }
 
 type StringAnnotation struct {